ingest/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
asyncpg>=0.29.0
boto3>=1.34.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pytz>=2024.1

